Addressing Alex comments
