+++ rundeckapp/grails-app/conf/spring/resources.groovy
@@ -888,8 +888,7 @@ beans={
             [it.key.toString(), it.value.toString()]
         }
         useForwardHeaders = useForwardHeadersConfig ?: Boolean.getBoolean('rundeck.jetty.connector.forwarded')
-        serverUrl = grailsApplication.config.getProperty('server.address', String.class)
-        serverPort = grailsApplication.config.getProperty('server.port', String.class)
+        serverUrl = grailsApplication.config.getProperty('grails.serverURL', String.class)
     }
 
     def stsMaxAgeSeconds = grailsApplication.config.getProperty("rundeck.web.jetty.servlet.stsMaxAgeSeconds",Integer.class,-1)
